--- src/app/projects/category/[type]/constants.ts
@@ -1,32 +1,0 @@
-import { v4 as uuid } from "uuid";
-
-export const stateProjects = [
-  {
-    id: uuid(),
-    project_state: "Lagos State",
-    country: "Nigeria",
-    no_of_farms: 50,
-    available_credits: 45000,
-  },
-  {
-    id: uuid(),
-    project_state: "Kano State",
-    country: "Nigeria",
-    no_of_farms: 50,
-    available_credits: 50000,
-  },
-  {
-    id: uuid(),
-    project_state: "Ogun State",
-    country: "Nigeria",
-    no_of_farms: 50,
-    available_credits: 50000,
-  },
-  {
-    id: uuid(),
-    project_state: "Kebbi State",
-    country: "Nigeria",
-    no_of_farms: 50,
-    available_credits: 50000,
-  },
-];
