+++ src/components/PaystackRedirection/types.ts
@@ -1,0 +1,4 @@
+export interface PaystackRedirectionProps {
+    type: "success" | "error";
+  }
+
